seperation of containers from their parent component to one file
