+++ src/components/Container.js
@@ -35,6 +35,30 @@ export const Padding = styled(Container)`
   }
 `;
 
+export const InfoContainer = styled(Padding)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr 2fr;
+  }
+  @media screen and (max-width: ${breakpoints.mobileLScreen}) {
+    grid-template-rows: 200px 1fr;
+  }
+`;
+
+export const SkillsContainer = styled(Padding)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    grid-template-rows: 1fr 300px 200px;
+    grid-template-columns: 1fr;
+  }
+
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    grid-template-rows: 1fr 1fr 200px;
+  }
+`;
+
 export const ContainerSkillsStack = styled(Container)`
   background-color: ${(props) => props.theme.color.backgroundSkillsStack};
   border-radius: ${(props) => props.theme.borderRadius};
@@ -44,7 +68,37 @@ export const ContainerSkillsStackItem = styled(Container)`
   background-color: ${(props) => props.theme.color.backgroundSkillsStack};
 `;
 
+export const ExperienceContainer = styled(Padding)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    grid-template-rows: 800px;
+    grid-template-columns: 2fr 1fr;
+  }
+  @media screen and (max-width: ${breakpoints.laptopLScreen}) {
+    grid-template-rows: 1fr;
+  }
+
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    grid-template-rows: 1fr 500px;
+    grid-template-columns: 1fr;
+  }
+`;
+
 export const ContainerExperienceYearsItem = styled(Container)`
   border: 4px solid ${(props) => props.theme.color.text};
   border-radius: 4px;
+`;
+
+export const ProjectsContainer = styled(Padding)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr 2fr 100px;
+  }
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    grid-template-columns: 1fr;
+    grid-template-rows: 700px 1fr 100px;
+  }
+  @media screen and (max-width: ${breakpoints.tabletScreen}) {
+    grid-template-rows: 1fr 3fr 100px;
+    grid-template-columns: 1fr;
+  }
 `;
