+++ src/pages/Projects.js
@@ -1,26 +1,9 @@
 import React from "react";
-import { Container, Padding } from "../components/Container";
-import { breakpoints } from "./../helpers/Media";
+import { Container, ProjectsContainer } from "../components/Container";
 import { Text } from "./../components/Text";
-import styled from "styled-components";
 import Quotes from "../components/Quotes";
 import ProjectList from "../components/Projects/ProjectList";
 import { HashLink } from "react-router-hash-link";
-
-const ProjectsContainer = styled(Padding)`
-  @media screen and (max-width: ${breakpoints.screen4k}) {
-    grid-template-columns: 1fr;
-    grid-template-rows: 1fr 2fr 100px;
-  }
-  @media screen and (max-width: ${breakpoints.laptopScreen}) {
-    grid-template-columns: 1fr;
-    grid-template-rows: 700px 1fr 100px;
-  }
-  @media screen and (max-width: ${breakpoints.tabletScreen}) {
-    grid-template-rows: 1fr 3fr 100px;
-    grid-template-columns: 1fr;
-  }
-`;
 
 export default function Projects(props) {
   return (
